requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
psycopg2-binary
pyjwt==2.8.0
passlib[bcrypt]==1.7.4
authlib==1.3.0
python-multipart==0.0.6
python-dotenv==1.0.0